server/requirements.txt
flask
flask-sqlalchemy
flask-cors
firebase-admin
python-dotenv
Werkzeug
requests
PyJWT
Pillow
openai
psycopg2-binary
